Warnings Mesh_3 regarding non existing file (documentation)
cherry-picked from 854c63f16bbb81fdb93145f9a8aa363753e61bf4

Currently we get a number of warnings when building the documentation for Mesh_3:
```
/home/cgal-testsuite/cgal_doc_build/CGAL-5.3.1-I-169/doc/Mesh_3/Mesh_3.txt:743: warning: unable to resolve reference to 'CGAL/Mesh_3/initialize_triangulation_from_labeled_image.h\' for \ref command
/home/cgal-testsuite/cgal_doc_build/CGAL-5.3.1-I-169/doc/Mesh_3/Mesh_3.txt:801: warning: unable to resolve reference to 'Mesh_3/random_labeled_image.h\' for \ref command
/home/cgal-testsuite/cgal_doc_build/CGAL-5.3.1-I-169/doc/Mesh_3/Mesh_3.txt:817: warning: unable to resolve reference to 'CGAL/Mesh_3/initialize_triangulation_from_gray_image.h\' for \ref command
```
Note the backslash at the end of the file name.

Most likely an attempt was made to escape the dot to not have it as part of the filename, though at least in doxygen 1.8.13 a closing dot was not seen as part of the filename.
Looking at the other escape character of doxygen: the commercial at sign, this was apparently working as escape in the filename as well, though in doxygen 1.8.15 this was not the case anymore either.
During one of the last changes also the backslash lost this possibility for this case and this is now consistent.

The current change (removing the backslashes) doesn't look like to alter the behavior for the 1.8.13 and master version in respect to CGAL.