Bugfix: ORIGINAL_CMAKE_MODULE_PATH was set to PARENT_SCOPE and later referenced in the current scope
In the `cgal_setup_module_path` function, the intent is to set a few variables then later export them.
The ORIGINAL_CMAKE_MODULE_PATH variable, however, was set directly in the parent scope. When later trying to export it, it reads from the current scope, where the variable is empty as it hasn't yet been set.
This fix makes correctly sets the variable in the local scope the first time, making the export work as expected.