Fix a use-after-free
In the expression `M[--L.end()] = M[start];`, the hash map `M` can be
rehashed during the evaluation of the expression. That can lead to a
use-after-free when the reference to `M[start]` is loaded, and the evaluation
of `M[--L.end()]` triggers a rehash, and invalidates the reference to
`M[start]`.