Fix the bug!
See the empty columns in https://cgal.geometryfactory.com/CGAL/testsuite/results-5.6-Ic-100.shtml
and the CMake output at
https://cgal.geometryfactory.com/CGAL/testsuite/CGAL-5.6-Ic-100/Installation/TestReport_Christo_MSVC-2022-Community-Release.gz

With `Scripts/developer_scripts/run_testsuite_with_ctest`, CMake is
called with `-DWITH_tests=ON -DCGAL_TEST_SUITE=ON`.

We do not want to disable that option `-DWITH_tests=ON` because it
is crucial for the correct behavior of `run_testsuite_with_ctest`.