Add support for REUSE
https://reuse.software/

The directories `.reuse/` and `LICENSES/` are for the Git layout.
The same directories in `Installation/` are for the tarball layout.

I have verified that with this patch `v5.5.2` and its release tarball
are compliant with the REUSE specifications.

The shady part is the file `.reuse/DEP5`, kind of equivalent to that
paragraph in `LICENSE.md`:
> A lot of files in the Git repository are not distributed in the source
> tarballs, for examples all the files in the `doc/` and `test/`
> sub-directories of CGAL packages. For all those files, unless they
> have an explicit license notice, the license is the
> Creative Commons CC0 1.0.

We should review the list of files corresponding to the pattern in that
file `.reuse/DEP5`.