Defining a new resource directory to be used with doxygen master 1.10.0
Doxygen master 1.10.0 has support for the c++ concept of modules but as in the past doxygen used the term "modules" for the grouping of information (`commands like `\defgroup`, `\ingroup` etc.) an new term for the later has been introduced: Topics.
The introduction of the new term also has some influence on the generated files and file names.
The consequence of this is that part of CGAL documentation (i.e. the "Reference Manual") item didn't appear anymore in a.o. the treeview.
- creating a new setup
- adding the "topics" to the layout files as in the past the "modules" were.
- adjusting the "hack.js" where some manipulation of the TreeView is done and this had "modules" coded into it.
- adjusting the "header_packages.html" as this used the name "modules.js"