Treeview missing due to missing setting in header.html
The PR #7946 was withdraw because the `$cookie` was not necessary anymore in the `header...`  files as it was part of the of the `$treeview`.
In `header_package.html` the `$treeview` has been disables and as such the `cookie.js` is not added/
Adding the `cookie.js` explicitly.

The effect was that there was a divider between the treeniew and the textual documentation but it was on the far left (and thus invisible), ith could be moved but after a refresh it was "gone" again.