Fix dangling reference in Construct_center_3(Circle_3)
This fixes the issue, but it could still be broken e.g. by a kernel
that is based on a CGAL kernel, but redefines the Sphere_3 and
with Sphere_3-functor that would not return a const& (or in that
case the custom kernel needs to define the appropriate functors
too...)

But so much breaks into this configuration, that it's out of scope
for this fix.